pyproject.toml
[tool.poetry]
name = "signaturizer3d"
version = "0.1.0"
description = "Signaturizers trained on 3D conformations of molcecules that are sensitive to the 3D geometry of molecules. The 3D Signaturizers are able to distinguish between stereoisomers."
authors = ["Aksel Lenes <aksel.lenes@irbbarcelona.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.1"
tqdm = "^4.66.1"
addict = "^2.4.0"
rdkit-pypi = "^2022.9.5"
scipy = "^1.11.3"
pandas = "^2.1.2"
scikit-learn = "^1.3.2"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
mypy = "^1.6.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
